fix: update e2e test imports for restructured module paths
