+++ packages/openclaw-plugin/public/index.html
@@ -30,10 +30,10 @@
     />
     <meta name="twitter:image" content="https://app.manifest.build/og-image.png" />
     <script src="/theme-init.js"></script>
-    <script type="module" crossorigin src="/assets/index-BfZr6VLP.js"></script>
+    <script type="module" crossorigin src="/assets/index-jCSzE4wA.js"></script>
     <link rel="modulepreload" crossorigin href="/assets/vendor-K8fEFSq9.js">
     <link rel="modulepreload" crossorigin href="/assets/auth-C2bQ9WcQ.js">
-    <link rel="stylesheet" crossorigin href="/assets/index-CJeyUrjL.css">
+    <link rel="stylesheet" crossorigin href="/assets/index-CbWrZHJK.css">
   </head>
   <body>
     <div id="root"></div>
